Refactor line information to use run length encoding
Rather than storing a line number for every byte in the chunk, now
instead uses RLE to encode the line information. Getting a line
thus requires a little extra work, but this is off the main code
path so is fine from a performance point of view. That is, it is
only triggered when errors/warnings occur.